New util tests, update PER
