+++ tests/utils/test_schedule.py
@@ -16,7 +16,7 @@ def inverted_schedule():
                        num_steps=10, invert=True)
 
 @pytest.fixture(scope='function')
-def exponential_decay_schedule():
+def exp_decay_schedule():
   yield ExpDecaySchedule(start=1.0, end=0.1, num_steps=1000)
 
 
@@ -42,12 +42,12 @@ def test_underflow_steps(inverted_schedule: LinearSchedule):
 
   assert val == inverted_schedule.min_val
 
-def test_exponential_decay(exp_decay_schedule: ExpDecaySchedule):
+def test_exp_decay(exp_decay_schedule: ExpDecaySchedule):
   assert exp_decay_schedule.value == exp_decay_schedule.start
 
-def test_exponential_decay_asymptote(exp_decay_schedule: ExpDecaySchedule):
+def test_exp_decay_asymptote(exp_decay_schedule: ExpDecaySchedule):
   val = None
-  for _ in range(exp_decay_schedule.num_steps * 10):
+  for _ in range(exp_decay_schedule.num_steps * 15):
     val = exp_decay_schedule.value
 
-  assert (val - exp_decay_schedule.end) < 1e-4
+  assert (val - exp_decay_schedule.end) < 1e-6
