+++ torchrl/utils/__init__.py
@@ -1,4 +1,4 @@
-from .misc import set_seeds, minibatch_generator, get_gym_spaces
+from .misc import set_seeds, minibatch_generator, get_gym_spaces, import_usr_dir
 from .sum_tree import SumTree
 from .schedule import LinearSchedule, ExpDecaySchedule
 from .nop import Nop
